Refactors for front-end config; adds dotenv-webpack
