+++ client/.env.example
@@ -1,0 +1,2 @@
+S3_ACCESS_KEY=FILL_ME_IN
+S3_SECRET=FILL_ME_IN
