+++ client/config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  s3Creds: {
+    accessKey: process.env.S3_ACCESS_KEY || '',
+    secret: process.env.S3_SECRET || '',
+  },
+};
