+++ package.json
@@ -65,6 +65,7 @@
     "axios": "^0.21.1",
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
+    "dotenv-webpack": "^7.0.3",
     "express": "^4.17.1",
     "firebase": "^9.0.0",
     "firebase-admin": "^9.11.1",
