+++ server/.env.example
@@ -18,5 +18,3 @@ FIREBASE_AUTH_URI=FILL_ME_IN
 FIREBASE_TOKEN_URI=FILL_ME_IN
 FIREBASE_AUTH_PROVIDER_URL=FILL_ME_IN
 FIREBASE_CLIENT_CERT_URL=FILL_ME_IN
-S3_ACCESS_KEY=FILL_ME_IN
-S3_SECRET=FILL_ME_IN
