+++ webpack.config.js
@@ -3,6 +3,7 @@
 const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const Dotenv = require('dotenv-webpack');
 
 const isProduction = process.env.NODE_ENV === 'production';
 
@@ -23,6 +24,8 @@ const config = {
     }),
 
     new MiniCssExtractPlugin(),
+
+    new Dotenv({ path: 'client/.env' }),
 
     // new webpack.EnvironmentPlugin(['NODE_ENV', 'DEBUG'])
